pyproject.toml
[tool.poetry]
name = "finviz"
version = "1.3.5"
description = "Unofficial API for finviz.com"
authors = ["Mario Stoev <bg.mstoev@gmail.com>", "James Bury <jabury@sympatico.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.9.3"
requests = "^2.25.1"
aiohttp = "^3.7.4"
lxml = "^4.6.3"
urllib3 = "^1.26.4"
user_agent = "^0.1.9"
cssselect = "^1.1.0"
tqdm = "^4.59.0"
tenacity = "^7.0.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
pytest = "^6.2.3"
python-dateutil = "^2.8.1"

[tool.pytest.ini_options]
minversion = "6.0"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::urllib3.exceptions.InsecureRequestWarning",
]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
